+++ src/main/java/omegadrive/util/RomHolder.java
@@ -13,14 +13,14 @@ public class RomHolder {
 
     private final static Logger LOG = LogHelper.getLogger(RomHolder.class.getSimpleName());
 
-    public static final RomHolder EMPTY_ROM = new RomHolder(new int[1]);
+    public static final RomHolder EMPTY_ROM = new RomHolder(new byte[1]);
 
     public final int baseSize;
     public final int size;
     public final int romMask;
-    public final int[] data;
+    public final byte[] data;
 
-    public RomHolder(int[] rom) {
+    public RomHolder(byte[] rom) {
         this.data = Util.getPaddedRom(rom);
         this.size = data.length;
         this.baseSize = rom.length;
